project/requirements.txt
flet==0.24.1
sqlite3
hashlib
datetime
reportlab==4.0.7
pandas==2.1.3
fpdf2==2.7.6